Use PoolType instead of a string. Fix related issues
