--- balancer-js/src/modules/pools/types.ts
@@ -1,10 +1,0 @@
-export enum PoolType {
-  Weighted = 'Weighted',
-  Investment = 'Investment',
-  Stable = 'Stable',
-  MetaStable = 'MetaStable',
-  StablePhantom = 'StablePhantom',
-  LiquidityBootstrapping = 'LiquidityBootstrapping',
-  AaveLinear = 'AaveLinear',
-  ERC4626Linear = 'ERC4626Linear',
-}
