+++ balancer-js/src/modules/pricing/pricing.module.ts
@@ -1,5 +1,5 @@
 import { Swaps } from '@/modules/swaps/swaps.module';
-import { BalancerSdkConfig } from '@/types';
+import { BalancerSdkConfig, PoolType } from '@/types';
 import {
   SubgraphPoolBase,
   ZERO,
@@ -80,7 +80,7 @@ export class Pricing {
       );
       if (!poolData)
         throw new BalancerError(BalancerErrorCode.POOL_DOESNT_EXIST);
-      const pool = Pools.from(poolData.poolType);
+      const pool = Pools.from(poolData.poolType as PoolType);
       return pool.spotPriceCalculator.calcPoolSpotPrice(
         tokenIn,
         tokenOut,
